Revert "86'd the divider in the race/ethnicity chart."
