--- src/js/charts-sandbox/charts/cagov-chart-vaccination-groups-race-ethnicity-age/template.js
@@ -1,25 +1,0 @@
-import css from "./index.scss";
-
-export default function template(translationsObj) {
-  return /*html*/ `<div class="py-2">
-    <div class="container">
-      <div class="col-lg-12 bg-white py-4 ">
-
-        <div class="row">
-          <div class="col-lg-12 col-md-12 col-sm-12 mx-auto px-0">
-            <div class="chart-title">Vaccinations by Race / Ethnicity and Age</div>
-            <div class="svg-legend-holder"></div>
-          </div>
-        </div>
-
-        <div class="row re-race-ethnicity-age-chart-list"> </div>
-
-        <div class="row">
-          <div class="col-lg-9 col-md-9 col-sm-12 mx-auto px-0">
-            <p class="chart-data-label col-lg-10 mx-auto">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. </p>
-          </div>
-        </div>
-      </div><!--END col-12-->
-    </div><!--END CONTAINER-->
-  </div>`;
-}
