+++ src/js/charts-sandbox/index.js
@@ -1,8 +1,8 @@
 console.log("Loading charts-sandbox");
-import './charts/cagov-chart-vaccination-groups-race-ethnicity/index.js'
-import './charts/cagov-chart-vaccination-groups-gender/index.js'
-import './charts/cagov-chart-vaccination-groups-age/index.js'
-// import './charts/cagov-chart-vaccination-groups-race-ethnicity-age/index.js'
+import './charts/ve-race-ethnicity/index.js'
+import './charts/ve-gender/index.js'
+import './charts/ve-age/index.js'
+// import './charts/ve-race-ethnicity-age/index.js'
 import './search/index.js'
 import './county-buttons/buttons.js'
 import './filters.js'
