+++ src/js/vaccines/charts/cagov-chart-vaccination-groups-race-ethnicity/index.js
@@ -128,10 +128,9 @@ class CAGovVaccinationGroupsRaceEthnicity extends window.HTMLElement {
     rtlOverride(this); // quick fix for arabic
   }
 
-  // offset bottom two bars so we can add divider - temporarily turned off
+  // offset bottom two bars so we can add divider
   getYOffset(ci) {
-    // return ci < 7 ? 0 : 20;
-    return 0;
+    return ci < 7 ? 0 : 20;
   }
 
   getLegendText() {
@@ -174,16 +173,15 @@ class CAGovVaccinationGroupsRaceEthnicity extends window.HTMLElement {
   }
 
   renderExtras(svg, data, x, y) {
-    // horizontal divider turned off for now
-    // let group = svg.append("g");
-    // group
-    //   .append("rect")
-    //   .attr("fill", "#000000")
-    //   .attr("class", "divider")
-    //   .attr("y", y(6) + (this.bar_vspace * 7) / 12)
-    //   .attr("x", 0)
-    //   .attr("width", this.dimensions.width)
-    //   .attr("height", 0.75);
+    let group = svg.append("g");
+    group
+      .append("rect")
+      .attr("fill", "#000000")
+      .attr("class", "divider")
+      .attr("y", y(6) + (this.bar_vspace * 7) / 12)
+      .attr("x", 0)
+      .attr("width", this.dimensions.width)
+      .attr("height", 0.75);
   }
 
   listenForLocations() {
